make ui module and add more functionality and base classes
